Fix wrong memory addresses due to not accounted GROUPR offset
GROUPR writes gendf data in a compact way and let us know, what the offset is.

The dlayxs part of CCCR loops over the full energy group range in container variable `a` where it should loop only until the offset. Thereby, on chance uninitialized or invalid memory may be addressed and may lead at best to a crash or simply runs without error but resulting in values like 1.E+44 for delayed spectra.

This commit introduces the offset and corrects 2 loops. Moreover,  a check is introduced if the requested memory fits into the `a` variable. The `a` size variable `isiza` was moved to a parameter statement and renamed to account for 4 byte and 8 byte equivalence'd variables.